requirements.txt
# Core RAG Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# Embeddings & Vector Store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LLM Interface
ollama>=0.1.7

# Document Processing
pypdf2>=3.0.1
python-docx>=0.8.11
beautifulsoup4>=4.12.2
youtube-transcript-api>=0.6.1
pandas>=2.0.0

# CLI & Utilities
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Additional utilities
tiktoken>=0.5.0
numpy>=1.24.0
